created a homepage component and loaded it in the homepage which is a route in app.js
