+++ src/App.js
@@ -6,13 +6,12 @@ import {FontAwesomeIcon} from "@fortawesome/react-fontawesome"
 import {faSearch,faHeart,faEdit,faUser} from "@fortawesome/free-solid-svg-icons"
 
 import HomePage from '../src/pages/home'
+import NoteBook from './pages/notebook';
 
 
 class App extends Component {
   
   render(){
-    
-
     return (
       <div className="App">
        
@@ -20,6 +19,9 @@ class App extends Component {
        <Switch>
           <Route path='/' render={(props) => (
             <HomePage {...props}/>
+          )}/>
+          <Route path='/notebook' render={(props) => (
+            <NoteBook {...props}/>
           )}/>
           
           {/* and in case nothing matches, we redirect: */}
@@ -30,7 +32,7 @@ class App extends Component {
         <Navbar bg="light" expand="lg" fixed='bottom'>
           <Container fluid>
           <Nav.Link href="/"><div className='bottomNavElement'><FontAwesomeIcon icon={faSearch} size="2x"/>Search</div></Nav.Link>
-          <Nav.Link href="#notebook"><div className='bottomNavElement'><div className='notebook'><FontAwesomeIcon icon={faEdit} size="2x"/></div>Notebook</div></Nav.Link>
+          <Nav.Link href="/notebook"><div className='bottomNavElement'><div className='notebook'><FontAwesomeIcon icon={faEdit} size="2x"/></div>Notebook</div></Nav.Link>
           <Nav.Link href="#linkSaved"><div className='bottomNavElement'><FontAwesomeIcon icon={faHeart} size="2x"/>Saved</div></Nav.Link>
           <Nav.Link href="#linkProfile"><div className='bottomNavElement'><FontAwesomeIcon icon={faUser} size="2x"/>Profile</div></Nav.Link>
           </Container>
