+++ src/pages/notebook.jsx
@@ -1,0 +1,16 @@
+import { Component } from 'react';
+import '../App';
+
+class NoteBook extends Component {
+
+
+    render() {
+        return(
+            <>
+          
+            </>
+        );
+    }
+}
+
+export default NoteBook;
